requirements.txt
Flask==2.3.2
numpy==1.24.3
scikit-learn==1.3.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0